% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_from_data_gov.R
\name{download_temp_visa_holders_data_gov}
\alias{download_temp_visa_holders_data_gov}
\title{Download temp visa holders}
\usage{
download_temp_visa_holders_data_gov(
  search_term = "name:temporary visa",
  path = "data/DoHA_temp_visa_holders.xlsx"
)
}
\arguments{
\item{search_term}{(default = "name:temporary visa", character) the CKAN search term you are using to look for}

\item{path}{}
}
\description{
Download temp visa holders
}
